% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimacion_porcentaje.R
\name{estimacion_porcentaje}
\alias{estimacion_porcentaje}
\title{Estimacion de porcentajes}
\usage{
estimacion_porcentaje(tabla, variable, prefijo, peso_final)
}
\arguments{
\item{tabla}{data.frame con datos de Excale o Planea.}

\item{variable}{Variable de agrupacion}

\item{prefijo}{Prefijo de los pesos muestrales replicados}

\item{peso_final}{Peso muestral combinado}
}
\description{
Estima porcentajes con su error estandar e intervalos de confianza al 95%.
}
\author{
Juan Bosco Mendoza Vega
}

