% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_media.R
\encoding{UTF-8}
\name{graf_media}
\alias{graf_media}
\title{Gráficas de medias-}
\usage{
graf_media(x)
}
\arguments{
\item{x}{Data frame. Resultados de las funciones media_pob o media_pv.}
}
\value{
Un objeto gráfico de tipo ggplot.
}
\description{
Genera gráficas para medias estimadas con las funciones media_pob y media_pv.
}
\details{
Esta función usa ggplot2 para generar gráficas sencillas con los
  resultados de las funciones media_pob y media_pv.
}
\examples{
# Data frame con resultados de Planea 2018 de Nayarit.
# (Las estimaciones pueden ser imprecisas por ser un fragmento de todos los
# datos, se presentan unicamente con fines demostrativos.)

p18_nay

# Estimamos la media de la variable RFAB y asignamos el resultado a
# media_rfab.
media_rfab <- media_pob(x = p18_nay, variable = "RFAB", w_final = "W_FSTUWT",
w_rep = "W_FSTR")

# Generamos una gráfica de la media estimada con graf_media.
graf_media(x = media_rfab)

# Estimamos la media de la variable MAT, que tiene valores plausibles, y
# asignamos el resultado a media_mat.
media_mat <- media_pv(x = p18_nay, variable = "MAT", w_final = "W_FSTUWT",
w_rep = "W_FSTR")

# Generamos una gráfica de la media estimada con graf_media, no hay
# diferencia si usamos el resultado de media_pob o media_pv.
graf_media(x = media_mat)

}
\author{
Juan Bosco Mendoza Vega
}
