% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_pocentaje.R
\name{graf_porcentaje}
\alias{graf_porcentaje}
\title{Grafica de porcentajes}
\usage{
graf_porcentaje(tabla_porcentaje)
}
\arguments{
\item{tabla_resultado}{Tabla generada con \strong{estimacion_porcentaje}.}
}
\value{
Un objeto grafico.
}
\description{
Genera una grafica a partir de una tabla creada con \strong{estimacion_porcentaje}.
}
\author{
Juan Bosco Mendoza Vega
}

