% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_pocentaje_h.R
\name{graf_porcentaje_h}
\alias{graf_porcentaje_h}
\title{Grafica de porcentajes, hprizontal}
\usage{
graf_porcentaje_h(tabla_porcentaje)
}
\arguments{
\item{tabla_resultado}{Tabla generada con \strong{estimacion_porcentaje}.}
}
\value{
Un objeto grafico, con orientacion horizontal.
}
\description{
Genera una grafica a partir de una tabla creada con \strong{estimacion_porcentaje}.
}
\author{
Juan Bosco Mendoza Vega
}
