% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_prop.R
\encoding{UTF-8}
\name{graf_prop}
\alias{graf_prop}
\title{Gráficas de proporciones.}
\usage{
graf_prop(x)
}
\arguments{
\item{x}{Data frame. Resultados de la funcion prop_pob.}
}
\value{
Un objeto gráfico de tipo ggplot.
}
\description{
Genera gráficas para proporciones estimadas con prop_pob.
}
\details{
Estas funciones usan ggplot2 para generar gráficas sencillas con los
  resultados de la función prop_pob.
}
\examples{
# Data frame con resultados de Planea 2018 de Nayarit.
# (Las estimaciones pueden ser imprecisas por ser un fragmento de todos los
# datos, se presentan unicamente con fines demostrativos.)

p18_nay

# Estimamos la proporción de la variable MARGINC y asignamos el resultado a
# prop_mar.
prop_mar <- prop_pob(x = p18_nay, variable = "MARGINC", w_final = "W_FSTUWT",
w_rep = "W_FSTR")

# Generamos una gráfica de la proporción estimada con graf_prop.
graf_prop(x = prop_mar)

}
\author{
Juan Bosco Mendoza Vega
}
