% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_resultado.R
\name{graf_resultado}
\alias{graf_resultado}
\title{Grafica de resultados}
\usage{
graf_resultado(tabla_resultado)
}
\arguments{
\item{tabla_resultado}{Tabla generada con \strong{media_poblacion} o
\strong{puntuacion_plausible}.}
}
\value{
Un objeto grafico.
}
\description{
Genera una grafica a partir de una tabla creada con media_poblacion. Funciona
con resultados de multiples grupos.
}
\author{
Juan Bosco Mendoza Vega
}

