% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_resultado_h.R
\name{graf_resultado_h}
\alias{graf_resultado_h}
\title{Grafica de resultados, horizontal}
\usage{
graf_resultado_h(tabla_resultado)
}
\arguments{
\item{tabla_resultado}{Tabla generada con \strong{media_poblacion} o
\strong{puntuacion_plausible}.}
}
\value{
Un objeto grafico con orientacion horizontal.
}
\description{
Genera una grafica a partir de una tabla creada con media_poblacion. Funciona
con resultados de multiples grupos.
}
\author{
Juan Bosco Mendoza Vega
}

