% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiples_variables.R
\name{multiples_variables}
\alias{multiples_variables}
\title{Estimacion de media poblacion para multiples variables}
\usage{
multiples_variables(tabla, variables, prefijo, peso_final, grupo = NULL)
}
\arguments{
\item{tabla}{Un data frame.}

\item{variables}{Un vector de texto con variables}

\item{prefijo}{Prefijo de los pesos muestrales replicados}

\item{peso_final}{Peso muestral combinado}

\item{grupo}{Variable de agrupacion}
}
\value{
Tabla con resultados
}
\description{
Estima medias poblacionales con error estandar e intervalo de confianza para
multiples variables. No estima para variables con puntuaciones plausibles.
}
\details{
Deberia funcionar con Excale y Planea por igual. Pero ya veremos
}
\author{
Juan Bosco Mendoza Vega
}
\references{
OECD (2009) PISA Data Analysis Manual: SPSS and SAS, Second Edition. OECD.
}

