% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimacion_porcentaje.R
\name{estimacion_porcentaje}
\alias{estimacion_porcentaje}
\title{Estimacion de porcentajes}
\usage{
estimacion_porcentaje(tabla, variable, prefijo, peso_final, grupo = NULL,
  cifras_decimales = 2)
}
\arguments{
\item{tabla}{Data frame u objeto coercionable a data.frame con datos de
Excale o Planea.}

\item{variable}{Cadena de texto con la variable de la que se estimaran
porcentajes.}

\item{prefijo}{Cadena de texto con el prefijo de las variables que contienen
los pesos muestrales replicados.}

\item{peso_final}{Cadena de texto con el nombre de la variable que contiene
el peso muestral combinado.}

\item{grupo}{Cadena de texto opcional con el nombre de la variable o
variables de agrupacion que se usaran para estimar porcentajes. Por
defecto, ninguna.}

\item{cifras_decimales}{Numero de cifras decimales que se mostraran en los
resultados.}
}
\value{
Una Data frame que contiene la estimacion de porcentaje para el
  argumento variable, la estimacion del error estandar asociado a cada
  porcentaje y los limites inferior y superior de su intervalo de confianza
  al 95%.
}
\description{
Estima porcentajes de variables con su error estandar e intervalos de
confianza al 95%.
}
\examples{
# Una Data frame con los resultados de Planea 2015
planea2015

# Porcentaje de Sexo ("SEXO") con cuatro cifras decimales
estimacion_porcentaje(tabla = planea2015, variable = "SEXO",
prefijo = "W_FSTR", peso_final = "W_FSTUWT", cifras_decimales = 4)

# Porcentaje de Sexo ("SEXO"), agrupado por servicio educativo ("SERV")
estimacion_porcentaje(tabla = planea2015, variable = "SEXO",
prefijo = "W_FSTR", peso_final = "W_FSTUWT", grupo = "SERV")

# Es posible pedir porcentajes por mas de un grupo, pero esto produce estimaciones con errores altos.
#' estimacion_porcentaje(tabla = planea2015, variable = "SEXO",
prefijo = "W_FSTR", peso_final = "W_FSTUWT", grupo = c("NOM_ENT", "SERV"))

}
\references{
OECD (2009) PISA Data Analysis Manual: SPSS and SAS, Second
Edition. OECD.
}
\author{
Juan Bosco Mendoza Vega
}
