% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_poblacion.R
\name{media_poblacion}
\alias{media_poblacion}
\title{Media poblacional}
\usage{
media_poblacion(tabla, variable, prefijo, peso_final, grupo = NULL)
}
\arguments{
\item{tabla}{Data frame con una tabla de datos de Excale o Planea}

\item{variable}{Variable a la que se estimara media poblacional}

\item{prefijo}{Prefijo de los pesos muestrales replicados}

\item{peso_final}{Peso muestral combinado}

\item{grupo}{Variable de agrupacion}
}
\description{
Estima la media poblacional con error estandar e intervalos de confianza de
de una variable. Es posible estimar medias por subpoblaciones usando una
variable de agrupacion.
}
\author{
Juan Bosco Mendoza Vega
}
\references{
OECD (2009) PISA Data Analysis Manual: SPSS and SAS, Second Edition. OECD.
}

